requirements.txt
alembic==1.0.10
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.7.4
botocore==1.10.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==6.7
coverage==4.5.3
cryptography==2.7
docutils==0.14
Flask==0.12.2
Flask-Limiter==1.0.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-UUID==0.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.17
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==3.0.1
limits==1.3
Mako==1.0.12
MarkupSafe==1.0
mock==3.0.5
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pycparser==2.19
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.6.2
pytest-cov==2.7.1
python-dateutil==2.7.2
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.3.4
urllib3==1.25.3
uuid==1.30
wcwidth==0.1.7
Werkzeug==0.14.1
zipp==0.5.1
